+++ lib/ex_pdfium/native.ex
@@ -55,7 +55,7 @@ defmodule ExPdfium.Native do
   def document_extract_text(_doc, _page_index), do: :erlang.nif_error(:nif_not_loaded)
   def document_extract_text_all(_doc), do: :erlang.nif_error(:nif_not_loaded)
   def document_text_segments(_doc, _page_index), do: :erlang.nif_error(:nif_not_loaded)
-  def document_text_chars(_doc, _page_index), do: :erlang.nif_error(:nif_not_loaded)
+  def document_text_chars(_doc, _page_index, _with_style), do: :erlang.nif_error(:nif_not_loaded)
 
   # `query` is binary; `match_case`/`whole_word` are booleans.
   def document_search_text(_doc, _page_index, _query, _match_case, _whole_word),